Remove business bank account core component
