+++ packages/webcomponents/src/ui-components/form/form-helpers/form-control-tooltip/readme.md
@@ -23,7 +23,6 @@
  - [form-control-select](../..)
  - [form-control-text](../..)
  - [justifi-additional-questions-form-step](../../../../components/business-forms/payment-provisioning/additional-questions)
- - [justifi-business-bank-account-form-step-core](../../../../components/business-forms/payment-provisioning/bank-account)
  - [justifi-business-owners-form-step](../../../../components/business-forms/payment-provisioning/business-owners)
  - [justifi-business-representative-form-inputs](../../../../components/business-forms/payment-provisioning/business-representative)
  - [justifi-legal-address-form-step](../../../../components/business-forms/payment-provisioning/legal-address-form)
@@ -44,7 +43,6 @@ graph TD;
   form-control-select --> form-control-tooltip
   form-control-text --> form-control-tooltip
   justifi-additional-questions-form-step --> form-control-tooltip
-  justifi-business-bank-account-form-step-core --> form-control-tooltip
   justifi-business-owners-form-step --> form-control-tooltip
   justifi-business-representative-form-inputs --> form-control-tooltip
   justifi-legal-address-form-step --> form-control-tooltip
